The goal of the system was to try and remove the elements from the structure holding ingredients that had no actual ingredients required. So that the player would only see and keep track of the ingredients they need to get for their target recipe. Tried messing around with data tables and creating arrays but these did not function well. Telisa has created a data structure containing maps, so that you can access each randomly chosen recipe which by default stores only the ingredients needed. I will build upon this to add the tracking to the screen and then implement this into the scoring system when we have more details about how cooked food will be scored. 
The goal of the system was to try and remove the elements from the structure holding ingredients that had no actual ingredients required. So that the player would only see and keep track of the ingredients they need to get for their target recipe. Tried messing around with data tables and creating arrays but these did not function well. Telisa has created a data structure containing maps, so that you can access each randomly chosen recipe which by default stores only the ingredients needed. I will build upon this to add the tracking to the screen and then implement this into the scoring system when we have more details about how cooked food will be scored.