Cloche gun is capable of storing the enemy ingredient type before destroying them, sending it to the widget. This will check to see if it matches any on the list and cycle down until the correct amount of ingredients has been achieved, in which case it will tick off the ingredient. Will need to make an inventory or editing existing inventory actor component to store enemy values, which can then be used to spawned finished ingredients at the games hub later on. 
Cloche gun is capable of storing the enemy ingredient type before destroying them, sending it to the widget. This will check to see if it matches any on the list and cycle down until the correct amount of ingredients has been achieved, in which case it will tick off the ingredient. Will need to make an inventory or editing existing inventory actor component to store enemy values, which can then be used to spawned finished ingredients at the games hub later on.