Beginning The Food Assembler
Cloche projectile will need to be updated. Score is lost when an incorrect ingredient is captured and not taken to the hub. This means that the recipe will need to define what happens to captured ingredients. The cooking pot will calculate the worst of the ingredient type that makes it in and use those to calculate the cooking score, which is added to the final score