Finally fixed the issue with the food assembler not counting ingredients
Finally fixed the issue with the food assembler not counting ingredients. It would correctly apply the score if there was more than one of an ingredient captured. This is because of a check at the start of the cull function, which would not perform the function if there was only one of an ingredient. I thought that it would still release the new array but it simply makes it empty instead. This has been corrected and the cooking score now functions