Added Rotation Variable
Adding arrow that's rotation can be set on construction so that the level designers can change which way the player is facing when they teleport. It doesn't work currently but I believe this is due to small changed needing to be made to the player controller.

Now that I have looked into it a bit more, it's not a setting on the player controller itself, but the rotation of the player controller needed to be set rather than the rotation of the player character itself