Updated
Attempted to make this work by removing timer from the player HUD and adding it when the player first teleports but realised this wasn't an effective method. Oscar let me know that visibility could be set for Widgets so this was changed to make the timer invisible on begin play, and then set to visible when the player teleports into the main level.