Added code from the cloche to the food assembler. Found issue with food assembly where the funnel was spawning uncooked ingredients with bacon as it's cooking type rather than null. Found that this was due to bacon being set as the default in AC health, Oscar will change this on his side as he is editing ingredient types, which should fix my issue when merged.
Added code from the cloche to the food assembler. Found issue with food assembly where the funnel was spawning uncooked ingredients with bacon as it's cooking type rather than null. Found that this was due to bacon being set as the default in AC health, Oscar will change this on his side as he is editing ingredient types, which should fix my issue when merged.